split out pass module, write some docs
Most of this commit is documentation for the frame context and recording
system. This details how resource synchronization is performed while
recording the output of the render graph to a command buffer, and
introduces the frame context (name borrowed from Granite) as a logical
grouping of resources used together during a frame.